Añadidos permisos a CRUD acciones

+++ views/action/actionShowAllView.php
@@ -21,31 +21,30 @@ class ActionShowAllView
     function render()
     {
         ?>
-        
-        <head>
-            <link rel="stylesheet" href="../css/styles.css"/>
-            <link rel="stylesheet" href="../css/table.css"/>
-        </head>
         <main role="main" class="margin-main ml-sm-auto px-4">
             <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-4 pb-2 mb-3">
-                <h1 class="h2" data-translate="Listado de acciones"></h1>
+                <h2 class="h2" data-translate="Listado de acciones"></h2>
                 <!-- Search -->
-                <form class="row" action='../controllers/actionController.php' method='POST'>
-                    <input type="text" class="form-control" id="search" name="search" data-translate="Texto a buscar">
-                    <button name="submit" type="submit" class="btn btn-primary" data-translate="Buscar"></button>
+                <form class="row" action='../controllers/actionController.php?action=search' method='POST'>
+                    <div class="col-10 pr-1">
+                        <input type="text" class="form-control" id="nombre" name="nombre" data-translate="Texto a buscar">
+                    </div>
+                    <div class="col-2 pl-0">
+                        <button name="submit" type="submit" class="btn btn-primary" data-translate="Buscar"></button>
+                    </div>
                 </form>
 
-                <?php //if (!empty($this->stringToSearch)): ?>
+                <?php if (!empty($this->stringToSearch)): ?>
                     <a class="btn btn-primary mr-1" role="button" href="../controllers/defaultController.php">
                         <p data-translate="Volver"></p>
                     </a>
-                <?php //else:
-                    //if (HavePermission("Action", "ADD")): ?>
+                <?php else:
+                    if (checkPermission("accion", "ADD")): ?>
                         <a class="btn btn-success" role="button" href="../controllers/actionController.php?action=add">
                             <span data-feather="plus"></span>
                             <p data-translate="Añadir acción"></p>
                         </a>
-                    <?php //endif; endif; ?>
+                    <?php endif; endif; ?>
             </div>
             <div class="table-responsive">
                 <table class="table table-striped table-sm">
@@ -65,18 +64,18 @@ class ActionShowAllView
                             <td><?php echo $action->getNombre() ?></td>
                             <td><?php echo $action->getDescripcion() ?></td>
                             <td class="row">
-                                <?php // if (HavePermission("Action", "SHOWCURRENT")) { ?>
+                                <?php if (checkPermission("accion", "SHOWCURRENT")) { ?>
                                     <a href="../controllers/actionController.php?action=show&id=<?php echo $action->getId() ?>">
                                         <span data-feather="eye"></span></a>
-                                <?php //}
-                                //if (HavePermission("Action", "EDIT")) { ?>
+                                <?php }
+                                if (checkPermission("accion", "EDIT")) { ?>
                                     <a href="../controllers/actionController.php?action=edit&id=<?php echo $action->getId() ?>">
                                         <span data-feather="edit"></span></a>
-                                <?php //}
-                                //if (HavePermission("Action", "DELETE")) { ?>
+                                <?php }
+                                if (checkPermission("accion", "DELETE")) { ?>
                                     <a href="../controllers/actionController.php?action=delete&id=<?php echo $action->getId() ?>">
                                         <span data-feather="trash-2"></span></a>
-                                <?php //} ?>
+                                <?php } ?>
                             </td>
                         </tr>
                     <?php endforeach; ?>
